configs/requirements.txt
python-dotenv>=1.0
requests>=2.32
loguru>=0.7
pydantic>=2.6
tenacity>=8.2
rich>=13.7
typer>=0.12
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
fpdf2>=2.7
openai>=1.40
langchain>=0.2
chromadb>=0.5
faiss-cpu>=1.8
slack_sdk>=3.27
notion-client>=2.2
streamlit>=1.36
sqlalchemy>=2.0
python-dateutil>=2.9


